requirements.txt
Flask==3.0.3
python-dotenv
gunicorn==20.1.0